Fix type defaulting warning

+++ src/Silk/Opaleye/TH/Table.hs
@@ -51,7 +51,7 @@ makeType = \case
       appendToName (Name (OccName occ) ns) s = OccName (occ ++ s) `Name` ns
       tvars :: [Name]
       tvars = map mkName . take (length vtys) $ tyvarNames
-      tyvarNames = [ x:(show i) | i <- [0..], x <- ['a'..'z'] ] -- a0, b0, ..., a1, b1, ...
+      tyvarNames = [ x:(show i) | i <- [(0::Int)..], x <- ['a'..'z'] ] -- a0, b0, ..., a1, b1, ...
       ttys :: [Type]
       ttys = map (\(_,_,tp) -> tp) vtys
       ttysO :: [Type]
